// backend/src/models/TodoItem.ts
export interface TodoItem {
  userId: string
  todoId: string
  createdAt: string
  name: string
  dueDate: string
  done: boolean
  attachmentUrl?: string
}

export interface TodoItemDTO {
  userId?: string
  todoId?: string
  createdAt?: string
  name?: string
  dueDate?: string
  done?: boolean
  attachmentUrl?: string
}